+++ main.tf
@@ -1,3 +1,7 @@
+locals {
+  dynamodb_table_arn = try(aws_dynamodb_table.this[0].arn, aws_dynamodb_table.autoscaled[0].arn, aws_dynamodb_table.autoscaled_gsi_ignore[0].arn, "")
+}
+
 resource "aws_dynamodb_table" "this" {
   count = var.create_table && !var.autoscaling_enabled ? 1 : 0
 
@@ -11,6 +15,11 @@ resource "aws_dynamodb_table" "this" {
   stream_view_type            = var.stream_view_type
   table_class                 = var.table_class
   deletion_protection_enabled = var.deletion_protection_enabled
+  region                      = var.region
+  restore_date_time           = var.restore_date_time
+  restore_source_name         = var.restore_source_name
+  restore_source_table_arn    = var.restore_source_table_arn
+  restore_to_latest_time      = var.restore_to_latest_time
 
   ttl {
     enabled        = var.ttl_enabled
@@ -18,7 +27,8 @@ resource "aws_dynamodb_table" "this" {
   }
 
   point_in_time_recovery {
-    enabled = var.point_in_time_recovery_enabled
+    enabled                 = var.point_in_time_recovery_enabled
+    recovery_period_in_days = var.point_in_time_recovery_period_in_days
   }
 
   dynamic "attribute" {
@@ -52,6 +62,24 @@ resource "aws_dynamodb_table" "this" {
       read_capacity      = lookup(global_secondary_index.value, "read_capacity", null)
       write_capacity     = lookup(global_secondary_index.value, "write_capacity", null)
       non_key_attributes = lookup(global_secondary_index.value, "non_key_attributes", null)
+
+      dynamic "on_demand_throughput" {
+        for_each = try([global_secondary_index.value.on_demand_throughput], [])
+
+        content {
+          max_read_request_units  = try(on_demand_throughput.value.max_read_request_units, null)
+          max_write_request_units = try(on_demand_throughput.value.max_write_request_units, null)
+        }
+      }
+
+      dynamic "warm_throughput" {
+        for_each = try([global_secondary_index.value.warm_throughput], [])
+
+        content {
+          read_units_per_second  = try(warm_throughput.value.read_units_per_second, null)
+          write_units_per_second = try(warm_throughput.value.write_units_per_second, null)
+        }
+      }
     }
   }
 
@@ -59,16 +87,75 @@ resource "aws_dynamodb_table" "this" {
     for_each = var.replica_regions
 
     content {
-      region_name            = replica.value.region_name
-      kms_key_arn            = lookup(replica.value, "kms_key_arn", null)
-      propagate_tags         = lookup(replica.value, "propagate_tags", null)
-      point_in_time_recovery = lookup(replica.value, "point_in_time_recovery", null)
+      region_name                 = replica.value.region_name
+      kms_key_arn                 = lookup(replica.value, "kms_key_arn", null)
+      propagate_tags              = lookup(replica.value, "propagate_tags", null)
+      point_in_time_recovery      = lookup(replica.value, "point_in_time_recovery", null)
+      deletion_protection_enabled = lookup(replica.value, "deletion_protection_enabled", null)
+      consistency_mode            = try(replica.value.consistency_mode, null)
     }
   }
 
   server_side_encryption {
     enabled     = var.server_side_encryption_enabled
     kms_key_arn = var.server_side_encryption_kms_key_arn
+  }
+
+  dynamic "import_table" {
+    for_each = length(var.import_table) > 0 ? [var.import_table] : []
+
+    content {
+      input_format           = import_table.value.input_format
+      input_compression_type = try(import_table.value.input_compression_type, null)
+
+      dynamic "input_format_options" {
+        for_each = try([import_table.value.input_format_options], [])
+
+        content {
+
+          dynamic "csv" {
+            for_each = try([input_format_options.value.csv], [])
+
+            content {
+              delimiter   = try(csv.value.delimiter, null)
+              header_list = try(csv.value.header_list, null)
+            }
+          }
+        }
+      }
+
+      s3_bucket_source {
+        bucket       = import_table.value.bucket
+        bucket_owner = try(import_table.value.bucket_owner, null)
+        key_prefix   = try(import_table.value.key_prefix, null)
+      }
+    }
+  }
+
+  dynamic "on_demand_throughput" {
+    for_each = length(var.on_demand_throughput) > 0 ? [var.on_demand_throughput] : []
+
+    content {
+      max_read_request_units  = try(on_demand_throughput.value.max_read_request_units, null)
+      max_write_request_units = try(on_demand_throughput.value.max_write_request_units, null)
+    }
+  }
+
+  dynamic "warm_throughput" {
+    for_each = length(var.warm_throughput) > 0 ? [var.warm_throughput] : []
+
+    content {
+      read_units_per_second  = try(warm_throughput.value.read_units_per_second, null)
+      write_units_per_second = try(warm_throughput.value.write_units_per_second, null)
+    }
+  }
+
+  dynamic "global_table_witness" {
+    for_each = var.global_table_witness != null ? [var.global_table_witness] : []
+
+    content {
+      region_name = global_table_witness.value.region_name
+    }
   }
 
   tags = merge(
@@ -98,6 +185,11 @@ resource "aws_dynamodb_table" "autoscaled" {
   stream_view_type            = var.stream_view_type
   table_class                 = var.table_class
   deletion_protection_enabled = var.deletion_protection_enabled
+  region                      = var.region
+  restore_date_time           = var.restore_date_time
+  restore_source_name         = var.restore_source_name
+  restore_source_table_arn    = var.restore_source_table_arn
+  restore_to_latest_time      = var.restore_to_latest_time
 
   ttl {
     enabled        = var.ttl_enabled
@@ -105,7 +197,8 @@ resource "aws_dynamodb_table" "autoscaled" {
   }
 
   point_in_time_recovery {
-    enabled = var.point_in_time_recovery_enabled
+    enabled                 = var.point_in_time_recovery_enabled
+    recovery_period_in_days = var.point_in_time_recovery_period_in_days
   }
 
   dynamic "attribute" {
@@ -139,6 +232,24 @@ resource "aws_dynamodb_table" "autoscaled" {
       read_capacity      = lookup(global_secondary_index.value, "read_capacity", null)
       write_capacity     = lookup(global_secondary_index.value, "write_capacity", null)
       non_key_attributes = lookup(global_secondary_index.value, "non_key_attributes", null)
+
+      dynamic "on_demand_throughput" {
+        for_each = try([global_secondary_index.value.on_demand_throughput], [])
+
+        content {
+          max_read_request_units  = try(on_demand_throughput.value.max_read_request_units, null)
+          max_write_request_units = try(on_demand_throughput.value.max_write_request_units, null)
+        }
+      }
+
+      dynamic "warm_throughput" {
+        for_each = try([global_secondary_index.value.warm_throughput], [])
+
+        content {
+          read_units_per_second  = try(warm_throughput.value.read_units_per_second, null)
+          write_units_per_second = try(warm_throughput.value.write_units_per_second, null)
+        }
+      }
     }
   }
 
@@ -150,12 +261,62 @@ resource "aws_dynamodb_table" "autoscaled" {
       kms_key_arn            = lookup(replica.value, "kms_key_arn", null)
       propagate_tags         = lookup(replica.value, "propagate_tags", null)
       point_in_time_recovery = lookup(replica.value, "point_in_time_recovery", null)
+      consistency_mode       = try(replica.value.consistency_mode, null)
     }
   }
 
   server_side_encryption {
     enabled     = var.server_side_encryption_enabled
     kms_key_arn = var.server_side_encryption_kms_key_arn
+  }
+
+  dynamic "import_table" {
+    for_each = length(var.import_table) > 0 ? [var.import_table] : []
+
+    content {
+      input_format           = import_table.value.input_format
+      input_compression_type = try(import_table.value.input_compression_type, null)
+
+      dynamic "input_format_options" {
+        for_each = try([import_table.value.input_format_options], [])
+
+        content {
+
+          dynamic "csv" {
+            for_each = try([input_format_options.value.csv], [])
+
+            content {
+              delimiter   = try(csv.value.delimiter, null)
+              header_list = try(csv.value.header_list, null)
+            }
+          }
+        }
+      }
+
+      s3_bucket_source {
+        bucket       = import_table.value.bucket
+        bucket_owner = try(import_table.value.bucket_owner, null)
+        key_prefix   = try(import_table.value.key_prefix, null)
+      }
+    }
+  }
+
+  dynamic "on_demand_throughput" {
+    for_each = length(var.on_demand_throughput) > 0 ? [var.on_demand_throughput] : []
+
+    content {
+      max_read_request_units  = try(on_demand_throughput.value.max_read_request_units, null)
+      max_write_request_units = try(on_demand_throughput.value.max_write_request_units, null)
+    }
+  }
+
+  dynamic "warm_throughput" {
+    for_each = length(var.warm_throughput) > 0 ? [var.warm_throughput] : []
+
+    content {
+      read_units_per_second  = try(warm_throughput.value.read_units_per_second, null)
+      write_units_per_second = try(warm_throughput.value.write_units_per_second, null)
+    }
   }
 
   tags = merge(
@@ -189,6 +350,11 @@ resource "aws_dynamodb_table" "autoscaled_gsi_ignore" {
   stream_view_type            = var.stream_view_type
   table_class                 = var.table_class
   deletion_protection_enabled = var.deletion_protection_enabled
+  region                      = var.region
+  restore_date_time           = var.restore_date_time
+  restore_source_name         = var.restore_source_name
+  restore_source_table_arn    = var.restore_source_table_arn
+  restore_to_latest_time      = var.restore_to_latest_time
 
   ttl {
     enabled        = var.ttl_enabled
@@ -196,7 +362,8 @@ resource "aws_dynamodb_table" "autoscaled_gsi_ignore" {
   }
 
   point_in_time_recovery {
-    enabled = var.point_in_time_recovery_enabled
+    enabled                 = var.point_in_time_recovery_enabled
+    recovery_period_in_days = var.point_in_time_recovery_period_in_days
   }
 
   dynamic "attribute" {
@@ -241,6 +408,7 @@ resource "aws_dynamodb_table" "autoscaled_gsi_ignore" {
       kms_key_arn            = lookup(replica.value, "kms_key_arn", null)
       propagate_tags         = lookup(replica.value, "propagate_tags", null)
       point_in_time_recovery = lookup(replica.value, "point_in_time_recovery", null)
+      consistency_mode       = try(replica.value.consistency_mode, null)
     }
   }
 
@@ -265,4 +433,12 @@ resource "aws_dynamodb_table" "autoscaled_gsi_ignore" {
   lifecycle {
     ignore_changes = [global_secondary_index, read_capacity, write_capacity]
   }
+}
+
+resource "aws_dynamodb_resource_policy" "this" {
+  count = var.create_table && var.resource_policy != null ? 1 : 0
+
+  region       = var.region
+  resource_arn = local.dynamodb_table_arn
+  policy       = replace(var.resource_policy, "__DYNAMODB_TABLE_ARN__", local.dynamodb_table_arn)
 }